pyproject.toml
[project]
name = "MadokaBot"
version = "0.1.0"
description = "MadokaBot"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[aiohttp]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-console>=0.9.0",
    "nonebot-plugin-status>=0.9.0",
    "nonebot-plugin-alconna>=0.60.3",
    "nonebot-plugin-htmlrender>=0.6.7",
    "nonebot-plugin-datastore>=1.3.1",
    "nonebot-plugin-localstore>=0.7.4",
    "sqlalchemy>=2.0.45",
    "Pillow>=10.2.0",
    "pytz>=2024.2",
    "numpy>=1.24.4",
    "nonebot-plugin-apscheduler>=0.5.0",
]

[project.optional-dependencies]
dev = [
    
]

[tool.nonebot]
plugin_dirs = ["madokabot"]
builtin_plugins = []


[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-htmlrender = ["nonebot_plugin_htmlrender"]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-status = ["nonebot_plugin_status"]
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]
nonebot-plugin-localstore = ["nonebot_plugin_localstore"]
